flask_app/requirements.txt
Flask
pandas
gunicorn
google-auth
google-api-python-client
google-auth-oauthlib
Boto3
azure-cosmos
pyodbc
flask_cors
pymongo